requirements.txt
PyJWT==2.8.0
python-dotenv==1.1.0
